# signal/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Signal API",
    "version": "1.0.0",
    "description": "Basic auth + pings API with JWT and role-based access"
  },
  "servers": [
    { "url": "http://localhost:3000" }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "TokenResponse": {
        "type": "object",
        "properties": { "token": { "type": "string" } },
        "required": ["token"]
      },
      "RegisterRequest": {
        "type": "object",
        "properties": {
          "email": { "type": "string" },
          "password": { "type": "string", "minLength": 8 }
        },
        "required": ["email", "password"]
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "email": { "type": "string" },
          "password": { "type": "string", "minLength": 8 }
        },
        "required": ["email", "password"]
      },
...
    }
  },
  "paths": {
...
    "/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Register new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/RegisterRequest" } }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/TokenResponse" } }
            }
          },
          "409": { "description": "Email already registered" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/LoginRequest" } }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/TokenResponse" } }
            }
          },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
...
  }
}